+++ offchain/crates/solana-cli/src/command/shreds/validator_client_rewards/init_holding.rs
@@ -54,12 +54,7 @@ impl InitHoldingCommand {
             );
         }
 
-        let dz_connection = self
-            .solana_payer_options
-            .connection_options
-            .clone()
-            .into_shred_subscription_connection();
-        let wallet = Wallet::try_new(self.solana_payer_options, Some(dz_connection))?;
+        let wallet = Wallet::try_new(self.solana_payer_options, None)?;
         let wallet_key = wallet.pubkey();
 
         let vcr_key = find_validator_client_rewards_address(self.client_id).0;