+++ offchain/crates/solana-cli/src/command/shreds/withdraw.rs
@@ -50,12 +50,7 @@ pub struct WithdrawCommand {
 impl WithdrawCommand {
     pub async fn try_into_execute(self, dz_ledger_url: Option<String>) -> Result<()> {
         let moniker_env = self.solana_payer_options.connection_options.moniker_env();
-        let dz_connection = self
-            .solana_payer_options
-            .connection_options
-            .clone()
-            .into_shred_subscription_connection();
-        let wallet = Wallet::try_new(self.solana_payer_options, Some(dz_connection))?;
+        let wallet = Wallet::try_new(self.solana_payer_options, None)?;
         let wallet_key = wallet.pubkey();
 
         println!("Shred subscription - Withdraw (Close Payment Escrow)");